# use indexmap instead of linked hashmap, switch hasher to use ahash