--- a/src/langserver/completion.rs
+++ b/src/langserver/completion.rs
@@ -11,6 +11,8 @@ use dm::objtree::{TypeRef, TypeVar, TypeProc, ProcValue};
 use crate::{Engine, Span, is_constructor_name};
 use crate::symbol_search::contains;
 
+use ahash::RandomState;
+
 static PROC_KEYWORDS: &[&str] = &[
     // Implicit variables
     "args",
@@ -100,7 +102,7 @@ fn item_documentation(docs: &dm::docs::DocCollection) -> Option<Documentation> {
 
 fn items_ty<'a>(
     results: &mut Vec<CompletionItem>,
-    skip: &mut HashSet<(&str, &'a String)>,
+    skip: &mut HashSet<(&str, &'a String), RandomState>,
     ty: TypeRef<'a>,
     query: &str,
 ) {
@@ -253,7 +255,7 @@ impl<'a> Engine<'a> {
         }
 
         let mut next = Some(ty).filter(|ty| !ty.is_root());
-        let mut skip = HashSet::new();
+        let mut skip = HashSet::with_hasher(RandomState::default());
         while let Some(ty) = next {
             // override a parent's var
             for (name, var) in ty.get().vars.iter() {
@@ -347,7 +349,7 @@ impl<'a> Engine<'a> {
                 proc: None,
             }) => {
                 let mut next = Some(ty);
-                let mut skip = HashSet::new();
+                let mut skip = HashSet::with_hasher(RandomState::default());
                 while let Some(ty) = next {
                     // reference a declared proc
                     for (name, proc) in ty.get().procs.iter() {
@@ -443,7 +445,7 @@ impl<'a> Engine<'a> {
 
         // fields
         let mut next = Some(ty);
-        let mut skip = HashSet::new();
+        let mut skip = HashSet::with_hasher(RandomState::default());
         while let Some(ty) = next {
             items_ty(results, &mut skip, ty, query);
             next = ty.parent_type();
@@ -460,7 +462,7 @@ impl<'a> Engine<'a> {
         I: Iterator<Item = (Span, &'b Annotation)> + Clone,
     {
         let mut next = self.find_scoped_type(iter, priors);
-        let mut skip = HashSet::new();
+        let mut skip = HashSet::with_hasher(RandomState::default());
         while let Some(ty) = next {
             items_ty(results, &mut skip, ty, query);
             next = ty.parent_type_without_root();