--- a/src/langserver/document.rs
+++ b/src/langserver/document.rs
@@ -14,11 +14,13 @@ use lsp_types::{TextDocumentItem, TextDocumentIdentifier,
 
 use super::{invalid_request, url_to_path};
 
+use ahash::RandomState;
+
 /// A store for the contents of currently-open documents, with appropriate
 /// fallback for documents which are not currently open.
 #[derive(Default)]
 pub struct DocumentStore {
-    map: HashMap<Url, Document>,
+    map: HashMap<Url, Document, RandomState>,
 }
 
 impl DocumentStore {