--- a/src/tools/lib.rs
+++ b/src/tools/lib.rs
@@ -8,7 +8,6 @@ extern crate lodepng;
 extern crate inflate;
 
 #[macro_use] extern crate bytemuck;
-extern crate linked_hash_map;
 extern crate rand;
 extern crate bumpalo;