ahash is straight up faster for string hashing
indexmap is at least maintained and has methods that we use, also faster iteration.

before:
![before](https://user-images.githubusercontent.com/56778689/139512534-2e9b3a48-25c9-46bf-ba14-972b904d012d.png)
after:
![image](https://user-images.githubusercontent.com/56778689/139512587-b821d0c6-f99f-4bae-ad53-b5adbf65dc9a.png)