@SpaceManiac 2021-11-06
I saw a much smaller speedup on my machine, over several trials: dreamchecker average time from 5.946s to 5.794s, dm-langserver average time from 6.488s to 6.344s. Not nothing, but not as drastic as you reported. Can you try doing a few extra trials so we can find out if it your numbers were a fluke (disk cache can be a big confounder), or if there's a difference between our machines?