@jupyterkat 2021-11-07
might be because your machine just doesn’t support AES-NI like most modern machines do. have you tried on larger codebases where the extra seconds matter more