@SpaceManiac 2021-11-07
As far as I can tell my processor does support AES-NI. I was testing on /tg/station.